Moved input handler code to game_render.c
